Add VectorMap
This is effectively an object pool backed by a deque.
New objects are created by adding to the end of the deque.
Objects that are destroyed leave holes
which are tracked and reused for new objects.
IDs contain a generation number
so that replacement objects filling the holes
are not conflated with the previous object that was there.