Use old default log creation behaviour if no log arg passed
On Windows, RWE will fail to open the log file
if another instance is already running.
So we need a fallback behaviour in order for RWE to start up.

This is needed to e.g. test network game behaviour from the launcher,
where you want to start two instances of RWE on the same machine
but can't pass the --log argument.