use separate network send and receive buffers
This change is essentially out paranoia.
We send and receive independently,
so if we use the same buffer for sending and receiving,
we could start an async_receive_from, then do a bunch of send_to's,
then finally get a receive callback.

Did boost care that we used the buffer for some send_to's
in between giving it to async_receive_from
and our callback being fired?
Maybe not. Maybe it writes data into the buffer all at once,
immediately before giving it to us.
But then again, maybe it doesn't?

Using two buffers means that a buffer is only ever in use
for one operation at a time.