Implement ballistic projectiles
A difficult problem during implementation here was
compensating for overshooting in ballistics calculations.

Overshooting occurs because ballistics calculations
are performed assuming the projectile is fired
from the AimFrom location.
However, the projectile is actually fired from the
Query location (firing point).
Therefore the projectile ends up offset from its target
by the distance between these two points.

It seems that the developers of TA were aware of this
and did something to compensate for it.
From my testing, it appears that TA considers the Z distance
between the unit's aiming and firing points
when doing ballistic calculations.
I found that changing the Z distance between the aiming and firing point
made no visible difference to where a ballistic projectile would land.
Changing the X or Y distance however did influence
where the projectile would land.

It seems that the measurement of this Z distance
is done immediately after the call to Create().
I had a unit start with the barrel pointing up,
then had Create() start slowly rotating it back to horizontal.
When the unit was later ordered to fire,
the projectiles overshot their target.

In another test, I did the same rotation in Create(), but with a
"rotate now", so that the rotation happened immediately.
When the unit was later ordered to fire,
the projectiles did *not* overshoot their target.

Delete