Treat colored and textured vertices uniformly when rendering units
By having a uniform representation for unit vertices,
we can render units more efficiently
and reduce the number of cases we need to consider
in rendering code.

Immediate benefit of this is that colored faces are now drawn
when the unit is being built.
They weren't drawn previously because the shader ignored
colored faces.