Move test files into same tree as source
It was getting easy to forget about tests or forget
to keep the tree structure in sync with the source tree.

We adopt a new structure here to put the tests closer to the sources
so that it is more likely they will be considered when working
on the source.

To avoid any possible file name clashes we change the suffix
to .test.cpp rather than _test.cpp.