Remove spdlog/fmt/ostr.h include
The build has started failing with an obscure linker error
related to fmt when building with MSVC 19.29.30040.0 in Release mode
on AppVeyor.

It seems to build fine in MSVC 19.29.30133.0 on my local machine.
However since I don't want to get into debugging linker errors
let's just avoid the use of fmt for now.