Delete UiCamera, replace with viewport
We want to be able to dynamically change resolution
and these old cameras prevent that because they internally cache
the resolution.

We need the resolution to be held somewhere central
that we can easily update so the idea is to store it
in a central Viewport reference that is depended on
by the rest of the engine.