Move duplicated feature info out of MapFeature
This has a fairly noticable negative impact on performance
which I think is because of all the hash table lookups
we have to do now every frame to fetch feature definitions
and feature media infos. I am prepared to take the hit
for the time being for the sake of better architecture.

I hope we can eventually recover some of this performance
by storing these definitions in big arrays
with lookup by index instead.

We also significantly increase loading time and memory use
because we now eagerly pre-load every single feature's assets.
I hope that we can rework this soon so we only load features
that are used in the map and by units.