Remove fields from Unit in favour of UnitDefinition
Since UnitDefinition now holds lots of the fields
that used to live on Unit, lots of places now need to load
the UnitDefinition as well as the Unit in order to do their work.

Unit can now be thought of as the "state" of the unit,
while UnitDefinition holds the static attributes of that unit type.