Remove registry-js
It was too hard to keep this working while upgrading.
It is a native module that requires rebuilding
and some hacks to get it copied and wired into electron
when built.

This means that registry functionality will be broken for now
but we might be able to replace this with a home-grown
shell wrapper implementation later.