Don't assume units exist when processing events
Units are removed from the simulation on the same tick
that they died on.

This means that events that fired for those units on that tick
may not be actionable because we need to look up info
from the unit, but the unit also died that tick
and therefore no longer exists.