Restructure feature collision to better match original game
In the original game, features cannot overlap.
Some maps have features that overlap, and when these maps are loaded
one feature will "win" and actually get to be on the map,
while the rest don't appear.

It is also well known that corpses cannot spawn on top
of metal spots, which is another consequence of this behaviour.

RWE previously permitted features to overlap
but now it does not.